🚧 WIP finish implementing updating and reading goal angle and presentAngle with sync read. However, the robot does not move. Must debug.

+++ Dynamixel/ControlRegister.cs
@@ -4,7 +4,7 @@ public record ControlRegister(
     string Name,
     string Area,
     ushort Address,
-    int SizeInBytes,
+    ushort SizeInBytes,
     string Access,
     int? InitialValue)
 {
